backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.28.0
pdfplumber==0.11.0
python-docx==1.1.2
aiofiles==23.2.1
httpx==0.27.0
openpyxl==3.1.2
reportlab==4.2.0
python-pptx==0.6.23
